--- src/main/java/kitchenpos/delivery_order/application/DeliveryOrderService.java
@@ -1,157 +1,0 @@
-package kitchenpos.delivery_order.application;
-
-import kitchenpos.delivery_order.domain.DeliveryOrder;
-import kitchenpos.delivery_order.domain.DeliveryOrderRepository;
-import kitchenpos.delivery_order.domain.OrderStatus;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderType;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.*;
-
-@Service
-public class DeliveryOrderService {
-    private final DeliveryOrderRepository deliveryOrderRepository;
-    private final MenuRepository menuRepository;
-    private final KitchenridersClient kitchenridersClient;
-
-    public DeliveryOrderService(
-            final DeliveryOrderRepository deliveryOrderRepository,
-            final MenuRepository menuRepository,
-            final KitchenridersClient kitchenridersClient
-    ) {
-        this.deliveryOrderRepository = deliveryOrderRepository;
-        this.menuRepository = menuRepository;
-        this.kitchenridersClient = kitchenridersClient;
-    }
-
-    @Transactional
-    public DeliveryOrder create(final DeliveryOrder request) {
-        final OrderType type = request.getType();
-        if (Objects.isNull(type)) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItemRequests = request.getOrderLineItems();
-        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        final List<Menu> menus = menuRepository.findAllByIdIn(
-                orderLineItemRequests.stream()
-                        .map(OrderLineItem::getMenuId)
-                        .toList()
-        );
-        if (menus.size() != orderLineItemRequests.size()) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItems = new ArrayList<>();
-        for (final OrderLineItem orderLineItemRequest : orderLineItemRequests) {
-            final long quantity = orderLineItemRequest.getQuantity();
-            if (quantity < 0) {
-                throw new IllegalArgumentException();
-            }
-            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
-                    .orElseThrow(NoSuchElementException::new);
-            if (!menu.isDisplayed()) {
-                throw new IllegalStateException();
-            }
-            if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
-                throw new IllegalArgumentException();
-            }
-            final OrderLineItem orderLineItem = new OrderLineItem();
-            orderLineItem.setMenu(menu);
-            orderLineItem.setQuantity(quantity);
-            orderLineItems.add(orderLineItem);
-        }
-        DeliveryOrder order = new DeliveryOrder();
-        order.setId(UUID.randomUUID());
-        order.setType(type);
-        order.setStatus(OrderStatus.WAITING);
-        order.setOrderDateTime(LocalDateTime.now());
-        order.setOrderLineItems(orderLineItems);
-        final String deliveryAddress = request.getDeliveryAddress();
-        if (Objects.isNull(deliveryAddress) || deliveryAddress.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        order.setDeliveryAddress(deliveryAddress);
-        return deliveryOrderRepository.save(order);
-    }
-
-    @Transactional
-    public DeliveryOrder accept(final UUID orderId) {
-        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.WAITING) {
-            throw new IllegalStateException();
-        }
-        if (order.getType() == OrderType.DELIVERY) {
-            BigDecimal sum = BigDecimal.ZERO;
-            for (final OrderLineItem orderLineItem : order.getOrderLineItems()) {
-                sum = orderLineItem.getMenu()
-                        .getPrice()
-                        .multiply(BigDecimal.valueOf(orderLineItem.getQuantity()));
-            }
-            kitchenridersClient.requestDelivery(orderId, sum, order.getDeliveryAddress());
-        }
-        order.setStatus(OrderStatus.ACCEPTED);
-        return order;
-    }
-
-    @Transactional
-    public DeliveryOrder serve(final UUID orderId) {
-        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.ACCEPTED) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.SERVED);
-        return order;
-    }
-
-    @Transactional
-    public DeliveryOrder startDelivery(final UUID orderId) {
-        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getType() != OrderType.DELIVERY) {
-            throw new IllegalStateException();
-        }
-        if (order.getStatus() != OrderStatus.SERVED) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.DELIVERING);
-        return order;
-    }
-
-    @Transactional
-    public DeliveryOrder completeDelivery(final UUID orderId) {
-        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.DELIVERING) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.DELIVERED);
-        return order;
-    }
-
-    @Transactional
-    public DeliveryOrder complete(final UUID orderId) {
-        final DeliveryOrder order = deliveryOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        final OrderType type = order.getType();
-        final OrderStatus status = order.getStatus();
-        if (status != OrderStatus.DELIVERED) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.COMPLETED);
-        return order;
-    }
-
-    @Transactional(readOnly = true)
-    public List<DeliveryOrder> findAll() {
-        return deliveryOrderRepository.findAll();
-    }
-}
