--- src/main/java/kitchenpos/delivery_order/application/KitchenridersClient.java
@@ -1,8 +1,0 @@
-package kitchenpos.delivery_order.application;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-public interface KitchenridersClient {
-    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
-}
