--- src/main/java/kitchenpos/delivery_order/domain/DeliveryOrderRepository.java
@@ -1,14 +1,0 @@
-package kitchenpos.delivery_order.domain;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface DeliveryOrderRepository {
-    DeliveryOrder save(DeliveryOrder order);
-
-    Optional<DeliveryOrder> findById(UUID id);
-
-    List<DeliveryOrder> findAll();
-}
-
