--- src/main/java/kitchenpos/delivery_order/infra/DefaultKitchenridersClient.java
@@ -1,14 +1,0 @@
-package kitchenpos.delivery_order.infra;
-
-import kitchenpos.delivery_order.application.KitchenridersClient;
-import org.springframework.stereotype.Component;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-@Component
-public class DefaultKitchenridersClient implements KitchenridersClient {
-    @Override
-    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
-    }
-}
