--- src/main/java/kitchenpos/eat_in_order/application/EatInOrderService.java
@@ -1,133 +1,0 @@
-package kitchenpos.eat_in_order.application;
-
-import kitchenpos.eat_in_order.domain.*;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderType;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.*;
-
-@Service
-public class EatInOrderService {
-    private final EatInOrderRepository eatInOrderRepository;
-    private final MenuRepository menuRepository;
-    private final OrderTableRepository orderTableRepository;
-
-    public EatInOrderService(
-            final EatInOrderRepository eatInOrderRepository,
-            final MenuRepository menuRepository,
-            final OrderTableRepository orderTableRepository
-    ) {
-        this.eatInOrderRepository = eatInOrderRepository;
-        this.menuRepository = menuRepository;
-        this.orderTableRepository = orderTableRepository;
-    }
-
-    @Transactional
-    public EatInOrder create(final EatInOrder request) {
-        final OrderType type = request.getType();
-        if (Objects.isNull(type)) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItemRequests = request.getOrderLineItems();
-        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        final List<Menu> menus = menuRepository.findAllByIdIn(
-                orderLineItemRequests.stream()
-                        .map(OrderLineItem::getMenuId)
-                        .toList()
-        );
-        if (menus.size() != orderLineItemRequests.size()) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItems = new ArrayList<>();
-        for (final OrderLineItem orderLineItemRequest : orderLineItemRequests) {
-            final long quantity = orderLineItemRequest.getQuantity();
-            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
-                    .orElseThrow(NoSuchElementException::new);
-            if (!menu.isDisplayed()) {
-                throw new IllegalStateException();
-            }
-            if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
-                throw new IllegalArgumentException();
-            }
-            final OrderLineItem orderLineItem = new OrderLineItem();
-            orderLineItem.setMenu(menu);
-            orderLineItem.setQuantity(quantity);
-            orderLineItems.add(orderLineItem);
-        }
-        EatInOrder order = new EatInOrder();
-        order.setId(UUID.randomUUID());
-        order.setType(type);
-        order.setStatus(OrderStatus.WAITING);
-        order.setOrderDateTime(LocalDateTime.now());
-        order.setOrderLineItems(orderLineItems);
-
-        if (type == OrderType.EAT_IN) {
-            final OrderTable orderTable = orderTableRepository.findById(request.getOrderTableId())
-                    .orElseThrow(NoSuchElementException::new);
-            if (!orderTable.isOccupied()) {
-                throw new IllegalStateException();
-            }
-            order.setOrderTable(orderTable);
-        }
-        return eatInOrderRepository.save(order);
-    }
-
-    @Transactional
-    public EatInOrder accept(final UUID orderId) {
-        final EatInOrder order = eatInOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-
-        if (order.getStatus() != OrderStatus.WAITING) {
-            throw new IllegalStateException();
-        }
-
-        order.setStatus(OrderStatus.ACCEPTED);
-        return order;
-    }
-
-    @Transactional
-    public EatInOrder serve(final UUID orderId) {
-        final EatInOrder order = eatInOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.ACCEPTED) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.SERVED);
-        return order;
-    }
-
-    @Transactional
-    public EatInOrder complete(final UUID orderId) {
-        final EatInOrder order = eatInOrderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        final OrderStatus status = order.getStatus();
-
-        if (status != OrderStatus.SERVED) {
-            throw new IllegalStateException();
-        }
-
-        order.setStatus(OrderStatus.COMPLETED);
-        final OrderTable orderTable = order.getOrderTable();
-        if (!eatInOrderRepository.existsByOrderTableAndStatusNot(orderTable, OrderStatus.COMPLETED)) {
-            orderTable.setNumberOfGuests(0);
-            orderTable.setOccupied(false);
-        }
-
-        return order;
-    }
-
-    @Transactional(readOnly = true)
-    public List<EatInOrder> findAll() {
-        return eatInOrderRepository.findAll();
-    }
-}
