--- src/main/java/kitchenpos/eat_in_order/domain/OrderStatus.java
@@ -1,5 +1,0 @@
-package kitchenpos.eat_in_order.domain;
-
-public enum OrderStatus {
-    WAITING, ACCEPTED, SERVED, COMPLETED
-}
