+++ src/main/java/kitchenpos/order/delivery_order/application/KitchenridersClient.java
@@ -1,0 +1,8 @@
+package kitchenpos.order.delivery_order.application;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+public interface KitchenridersClient {
+    void requestDelivery(UUID orderId, BigDecimal amount, String deliveryAddress);
+}
