+++ src/main/java/kitchenpos/order/eat_in_order/domain/EatInOrderRepository.java
@@ -1,0 +1,18 @@
+package kitchenpos.order.eat_in_order.domain;
+
+import kitchenpos.order_table.domain.OrderTable;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface EatInOrderRepository {
+    EatInOrder save(EatInOrder order);
+
+    Optional<EatInOrder> findById(UUID id);
+
+    List<EatInOrder> findAll();
+
+    boolean existsByOrderTableAndStatusNot(OrderTable orderTable, OrderStatus status);
+}
+
