+++ src/main/java/kitchenpos/order/eat_in_order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.eat_in_order.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
