+++ src/main/java/kitchenpos/order/take_out_order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package kitchenpos.order.take_out_order.domain;
+
+public enum OrderStatus {
+    WAITING, ACCEPTED, SERVED, COMPLETED
+}
