+++ src/main/java/kitchenpos/order/take_out_order/domain/TakeOutOrder.java
@@ -1,0 +1,80 @@
+package kitchenpos.order.take_out_order.domain;
+
+import jakarta.persistence.*;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderType;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+
+@Table(name = "take_out_orders")
+@Entity
+public class TakeOutOrder {
+    @Column(name = "id", columnDefinition = "binary(16)")
+    @Id
+    private UUID id;
+
+    @Column(name = "type", nullable = false, columnDefinition = "varchar(255)")
+    @Enumerated(EnumType.STRING)
+    private OrderType type;
+
+    @Column(name = "status", nullable = false, columnDefinition = "varchar(255)")
+    @Enumerated(EnumType.STRING)
+    private OrderStatus status;
+
+    @Column(name = "order_date_time", nullable = false)
+    private LocalDateTime orderDateTime;
+
+    @OneToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE})
+    @JoinColumn(
+            name = "take_out_order_id",
+            nullable = false,
+            columnDefinition = "binary(16)",
+            foreignKey = @ForeignKey(name = "fk_order_line_item_to_take_out_orders")
+    )
+    private List<OrderLineItem> orderLineItems;
+
+    public TakeOutOrder() {
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public void setId(final UUID id) {
+        this.id = id;
+    }
+
+    public OrderType getType() {
+        return type;
+    }
+
+    public void setType(final OrderType type) {
+        this.type = type;
+    }
+
+    public OrderStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(final OrderStatus status) {
+        this.status = status;
+    }
+
+    public LocalDateTime getOrderDateTime() {
+        return orderDateTime;
+    }
+
+    public void setOrderDateTime(final LocalDateTime orderDateTime) {
+        this.orderDateTime = orderDateTime;
+    }
+
+    public List<OrderLineItem> getOrderLineItems() {
+        return orderLineItems;
+    }
+
+    public void setOrderLineItems(final List<OrderLineItem> orderLineItems) {
+        this.orderLineItems = orderLineItems;
+    }
+}
