--- src/main/java/kitchenpos/take_out_order/application/TakeOutOrderService.java
@@ -1,116 +1,0 @@
-package kitchenpos.take_out_order.application;
-
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderType;
-import kitchenpos.take_out_order.domain.OrderStatus;
-import kitchenpos.take_out_order.domain.TakeOutOrder;
-import kitchenpos.take_out_order.domain.TakeOutOrderRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.*;
-
-@Service
-public class TakeOutOrderService {
-    private final TakeOutOrderRepository orderRepository;
-    private final MenuRepository menuRepository;
-
-    public TakeOutOrderService(
-            final TakeOutOrderRepository orderRepository,
-            final MenuRepository menuRepository
-    ) {
-        this.orderRepository = orderRepository;
-        this.menuRepository = menuRepository;
-    }
-
-    @Transactional
-    public TakeOutOrder create(final TakeOutOrder request) {
-        final OrderType type = request.getType();
-        if (Objects.isNull(type)) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItemRequests = request.getOrderLineItems();
-        if (Objects.isNull(orderLineItemRequests) || orderLineItemRequests.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        final List<Menu> menus = menuRepository.findAllByIdIn(
-                orderLineItemRequests.stream()
-                        .map(OrderLineItem::getMenuId)
-                        .toList()
-        );
-        if (menus.size() != orderLineItemRequests.size()) {
-            throw new IllegalArgumentException();
-        }
-        final List<OrderLineItem> orderLineItems = new ArrayList<>();
-        for (final OrderLineItem orderLineItemRequest : orderLineItemRequests) {
-            final long quantity = orderLineItemRequest.getQuantity();
-            if (quantity < 0) {
-                throw new IllegalArgumentException();
-            }
-            final Menu menu = menuRepository.findById(orderLineItemRequest.getMenuId())
-                    .orElseThrow(NoSuchElementException::new);
-            if (!menu.isDisplayed()) {
-                throw new IllegalStateException();
-            }
-            if (menu.getPrice().compareTo(orderLineItemRequest.getPrice()) != 0) {
-                throw new IllegalArgumentException();
-            }
-            final OrderLineItem orderLineItem = new OrderLineItem();
-            orderLineItem.setMenu(menu);
-            orderLineItem.setQuantity(quantity);
-            orderLineItems.add(orderLineItem);
-        }
-        TakeOutOrder order = new TakeOutOrder();
-        order.setId(UUID.randomUUID());
-        order.setType(type);
-        order.setStatus(OrderStatus.WAITING);
-        order.setOrderDateTime(LocalDateTime.now());
-        order.setOrderLineItems(orderLineItems);
-        return orderRepository.save(order);
-    }
-
-    @Transactional
-    public TakeOutOrder accept(final UUID orderId) {
-        final TakeOutOrder order = orderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.WAITING) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.ACCEPTED);
-        return order;
-    }
-
-    @Transactional
-    public TakeOutOrder serve(final UUID orderId) {
-        final TakeOutOrder order = orderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        if (order.getStatus() != OrderStatus.ACCEPTED) {
-            throw new IllegalStateException();
-        }
-        order.setStatus(OrderStatus.SERVED);
-        return order;
-    }
-
-    @Transactional
-    public TakeOutOrder complete(final UUID orderId) {
-        final TakeOutOrder order = orderRepository.findById(orderId)
-                .orElseThrow(NoSuchElementException::new);
-        final OrderType type = order.getType();
-        final OrderStatus status = order.getStatus();
-
-        if (status != OrderStatus.SERVED) {
-            throw new IllegalStateException();
-        }
-
-        order.setStatus(OrderStatus.COMPLETED);
-        return order;
-    }
-
-    @Transactional(readOnly = true)
-    public List<TakeOutOrder> findAll() {
-        return orderRepository.findAll();
-    }
-}
