--- src/test/java/kitchenpos/eat_in_order/service/EatInOrderServiceTest.java
@@ -1,224 +1,0 @@
-package kitchenpos.eat_in_order.service;
-
-import kitchenpos.eat_in_order.application.EatInOrderService;
-import kitchenpos.eat_in_order.domain.EatInOrder;
-import kitchenpos.eat_in_order.domain.OrderStatus;
-import kitchenpos.eat_in_order.infra.InMemoryEatInOrderRepository;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.menu.infra.InMemoryMenuRepository;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderType;
-import kitchenpos.order_table.domain.OrderTable;
-import kitchenpos.order_table.domain.OrderTableRepository;
-import kitchenpos.order_table.infra.InMemoryOrderTableRepository;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.*;
-
-import java.math.BigDecimal;
-import java.util.*;
-
-import static kitchenpos.Fixtures.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-
-class EatInOrderServiceTest {
-    private InMemoryEatInOrderRepository orderRepository;
-    private MenuRepository menuRepository;
-    private OrderTableRepository orderTableRepository;
-    private EatInOrderService orderService;
-
-    @BeforeEach
-    void setUp() {
-        orderRepository = new InMemoryEatInOrderRepository();
-        menuRepository = new InMemoryMenuRepository();
-        orderTableRepository = new InMemoryOrderTableRepository();
-        orderService = new EatInOrderService(orderRepository, menuRepository, orderTableRepository);
-    }
-
-
-    @DisplayName("1개 이상의 등록된 메뉴로 매장 주문을 등록할 수 있다.")
-    @Test
-    void createEatInOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final UUID orderTableId = orderTableRepository.save(orderTable(true, 4)).getId();
-        final EatInOrder expected = createOrderRequest(OrderType.EAT_IN, orderTableId, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        final EatInOrder actual = orderService.create(expected);
-        assertThat(actual).isNotNull();
-        assertAll(
-                () -> assertThat(actual.getId()).isNotNull(),
-                () -> assertThat(actual.getType()).isEqualTo(expected.getType()),
-                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.WAITING),
-                () -> assertThat(actual.getOrderDateTime()).isNotNull(),
-                () -> assertThat(actual.getOrderLineItems()).hasSize(1),
-                () -> assertThat(actual.getOrderTable().getId()).isEqualTo(expected.getOrderTableId())
-        );
-    }
-
-    @DisplayName("주문 유형이 올바르지 않으면 등록할 수 없다.")
-    @NullSource
-    @ParameterizedTest
-    void create(final OrderType type) {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final EatInOrder expected = createOrderRequest(type, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        assertThatThrownBy(() -> orderService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("메뉴가 없으면 등록할 수 없다.")
-    @MethodSource("orderLineItems")
-    @ParameterizedTest
-    void create(final List<OrderLineItem> orderLineItems) {
-        final EatInOrder expected = createOrderRequest(OrderType.EAT_IN, orderLineItems);
-        assertThatThrownBy(() -> orderService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    private static List<Arguments> orderLineItems() {
-        return Arrays.asList(
-                null,
-                Arguments.of(Collections.emptyList()),
-                Arguments.of(Arrays.asList(createOrderLineItemRequest(INVALID_ID, 19_000L, 3L)))
-        );
-    }
-
-    @DisplayName("매장 주문은 주문 항목의 수량이 0 미만일 수 있다.")
-    @ValueSource(longs = -1L)
-    @ParameterizedTest
-    void createEatInOrder(final long quantity) {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final UUID orderTableId = orderTableRepository.save(orderTable(true, 4)).getId();
-        final EatInOrder expected = createOrderRequest(
-                OrderType.EAT_IN, orderTableId, createOrderLineItemRequest(menuId, 19_000L, quantity)
-        );
-        assertDoesNotThrow(() -> orderService.create(expected));
-    }
-
-    @DisplayName("빈 테이블에는 매장 주문을 등록할 수 없다.")
-    @Test
-    void createEmptyTableEatInOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final UUID orderTableId = orderTableRepository.save(orderTable(false, 0)).getId();
-        final EatInOrder expected = createOrderRequest(
-                OrderType.EAT_IN, orderTableId, createOrderLineItemRequest(menuId, 19_000L, 3L)
-        );
-        assertThatThrownBy(() -> orderService.create(expected))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("숨겨진 메뉴는 주문할 수 없다.")
-    @Test
-    void createNotDisplayedMenuOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, false, menuProduct())).getId();
-        final EatInOrder expected = createOrderRequest(OrderType.EAT_IN, createOrderLineItemRequest(menuId, 19_000L, 3L));
-        assertThatThrownBy(() -> orderService.create(expected))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문한 메뉴의 가격은 실제 메뉴 가격과 일치해야 한다.")
-    @Test
-    void createNotMatchedMenuPriceOrder() {
-        final UUID menuId = menuRepository.save(menu(19_000L, true, menuProduct())).getId();
-        final EatInOrder expected = createOrderRequest(OrderType.EAT_IN, createOrderLineItemRequest(menuId, 16_000L, 3L));
-        assertThatThrownBy(() -> orderService.create(expected))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("주문을 접수한다.")
-    @Test
-    void accept() {
-        final UUID orderId = orderRepository.save(order(OrderStatus.WAITING, orderTable(true, 4))).getId();
-        final EatInOrder actual = orderService.accept(orderId);
-        assertThat(actual.getStatus()).isEqualTo(OrderStatus.ACCEPTED);
-    }
-
-    @DisplayName("접수 대기 중인 주문만 접수할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "WAITING", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void accept(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status, orderTable(true, 4))).getId();
-        assertThatThrownBy(() -> orderService.accept(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("포장 및 매장 주문의 경우 서빙된 주문만 완료할 수 있다.")
-    @EnumSource(value = OrderStatus.class, names = "SERVED", mode = EnumSource.Mode.EXCLUDE)
-    @ParameterizedTest
-    void completeTakeoutAndEatInOrder(final OrderStatus status) {
-        final UUID orderId = orderRepository.save(order(status)).getId();
-        assertThatThrownBy(() -> orderService.complete(orderId))
-                .isInstanceOf(IllegalStateException.class);
-    }
-
-    @DisplayName("주문 테이블의 모든 매장 주문이 완료되면 빈 테이블로 설정한다.")
-    @Test
-    void completeEatInOrder() {
-        final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
-        final EatInOrder expected = orderRepository.save(order(OrderStatus.SERVED, orderTable));
-        final EatInOrder actual = orderService.complete(expected.getId());
-        assertAll(
-                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.COMPLETED),
-                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().isOccupied()).isFalse(),
-                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().getNumberOfGuests()).isEqualTo(0)
-        );
-    }
-
-    @DisplayName("완료되지 않은 매장 주문이 있는 주문 테이블은 빈 테이블로 설정하지 않는다.")
-    @Test
-    void completeNotTable() {
-        final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
-        orderRepository.save(order(OrderStatus.ACCEPTED, orderTable));
-        final EatInOrder expected = orderRepository.save(order(OrderStatus.SERVED, orderTable));
-        final EatInOrder actual = orderService.complete(expected.getId());
-        assertAll(
-                () -> assertThat(actual.getStatus()).isEqualTo(OrderStatus.COMPLETED),
-                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().isOccupied()).isTrue(),
-                () -> assertThat(orderTableRepository.findById(orderTable.getId()).get().getNumberOfGuests()).isEqualTo(4)
-        );
-    }
-
-    @DisplayName("주문의 목록을 조회할 수 있다.")
-    @Test
-    void findAll() {
-        final OrderTable orderTable = orderTableRepository.save(orderTable(true, 4));
-        orderRepository.save(order(OrderStatus.SERVED, orderTable));
-        final List<EatInOrder> actual = orderService.findAll();
-        assertThat(actual).hasSize(1);
-    }
-
-    private EatInOrder createOrderRequest(final OrderType orderType, final OrderLineItem... orderLineItems) {
-        return createOrderRequest(orderType, Arrays.asList(orderLineItems));
-    }
-
-    private EatInOrder createOrderRequest(final OrderType orderType, final List<OrderLineItem> orderLineItems) {
-        final EatInOrder order = new EatInOrder();
-        order.setType(orderType);
-        order.setOrderLineItems(orderLineItems);
-        return order;
-    }
-
-    private EatInOrder createOrderRequest(
-            final OrderType type,
-            final UUID orderTableId,
-            final OrderLineItem... orderLineItems
-    ) {
-        final EatInOrder order = new EatInOrder();
-        order.setType(type);
-        order.setOrderTableId(orderTableId);
-        order.setOrderLineItems(Arrays.asList(orderLineItems));
-        return order;
-    }
-
-    private static OrderLineItem createOrderLineItemRequest(final UUID menuId, final long price, final long quantity) {
-        final OrderLineItem orderLineItem = new OrderLineItem();
-        orderLineItem.setSeq(new Random().nextLong());
-        orderLineItem.setMenuId(menuId);
-        orderLineItem.setPrice(BigDecimal.valueOf(price));
-        orderLineItem.setQuantity(quantity);
-        return orderLineItem;
-    }
-}
